requirements.txt
torch==2.0.1+cu118 
transformers==4.34.0
diffusers==0.24.0
accelerate==0.22.0
Pillow==10.0.1
requests==2.31.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
